requirements.txt
aiohttp==3.7.3
alabaster==0.7.12
altgraph==0.17
aniso8601==9.0.1
APScheduler==3.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
Flask==1.0.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-serializer==0.0.5.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.0.0
idna==2.8
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
kiwisolver==1.3.1
lxml==4.6.2
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.11.1
multidict==5.1.0
numpy==1.19.4
packaging==20.9
pefile==2019.4.18
pep8==1.7.1
Pillow==7.0.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
pyTelegramBotAPI==3.7.4
pytest==6.2.2
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.14.0
python-editor==1.0.4
python-pptx==0.6.18
python-telegram-bot==13.1
pytube==10.4.1
pytz==2020.4
pywin32-ctypes==0.2.0
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==2.1
SQLAlchemy==1.4.4
SQLAlchemy-serializer==1.3.4.4
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.24.3
vk-api==11.9.1
Werkzeug==1.0.1
wikipedia==1.4.0
WTForms==2.3.3
yarl==1.6.3